scheduler-service/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
pyyaml==6.0.2
redis==5.1.1
httpx==0.27.2
psutil==6.0.0
pyinstaller==6.16.0
structlog==22.1.0
websockets==12.0